add stream test
