+++ bcs-app/backend/tests/container_service/observability/log_stream/test_log_fetch.py
@@ -1,0 +1,47 @@
+# -*- coding: utf-8 -*-
+#
+# Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
+# Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
+# Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://opensource.org/licenses/MIT
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+# an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+#
+import os.path
+
+import pytest
+
+from backend.container_service.observability.log_stream import utils
+
+
+class TestLogStream:
+    @pytest.fixture
+    def log_content(self):
+        with open(os.path.join(os.path.dirname(__file__), "fake_log.txt"), 'rb') as f:
+            return f.read().decode('utf-8')
+
+    def test_fetch(self, api_client, project_id, cluster_id, namespace, pod_name, container_name):
+        """ 测试获取日志 """
+        response = api_client.get(
+            f'/api/logs/projects/{project_id}/clusters/{cluster_id}/namespaces/{namespace}/pods/{pod_name}/stdlogs/'
+        )
+        assert response.json()['code'] == 0
+
+    def test_refine_k8s_logs(self, log_content):
+        logs = utils.refine_k8s_logs(log_content, None)
+        assert len(logs) == 10
+        assert logs[0].time == '2021-05-19T12:03:52.516011121Z'
+
+    def test_calc_since_time(self, log_content):
+        logs = utils.refine_k8s_logs(log_content, None)
+        sine_time = utils.calc_since_time(logs[0].time, logs[-1].time)
+        assert sine_time == '2021-05-19T12:03:10.125788125Z'
+
+    def test_calc_previous_page(self, log_content):
+        logs = utils.refine_k8s_logs(log_content, None)
+        page = utils.calc_previous_page(logs, {'container_name': "", "previous": ""}, "")
+        assert page != ""
