+++ bcs-app/pyproject.toml
@@ -63,8 +63,9 @@ flake8 = "3.7.8"
 ipdb = "0.10.3"
 ipython = "6.1.0"
 mock = "3.0.5"
-pytest = "4.5.0"
+pytest = "6.2.4"
 pytest-django = "3.9.0"
+pytest-asyncio = "^0.15.1"
 
 [tool.black]
 exclude = '''
